requirements.txt
httpx
msgpack-python
numpy
pillow
pyoxipng
websocket-client
zstd
